Extract duplicate code to package-private JsonExceptionMappers
